games/tic-tac-toe/archive-service/requirements.txt
kafka-python==2.0.2
pymongo==4.6.1

